Add minimal domestication test
Instantiate two allocators and arrange for a message to get passed between them
by exploiting the existing slow-paths' handling of message queues.  Count and
CHECK the number of domestication calls during this message passing.  For a
little more excitement, pave over the forward pointer in the freelist::Object::T
that is the message and have the domestication callback patch the original value
back; should we somehow fail to invoke the domestication callback on that
address, this will induce a crash (WHP, on both CHECK_CLIENT and unchecked
builds).