malloc test: correctly abort test on unexpected errno.
Previously this test would emit an error message but not
abort if the errno was not as expected on failure. This
was because the return in the null == true case prevented the
check for failed == true at the end of check_result from
being reached. To resolve this just abort immediately instead
of instead of setting failed to true, as in the null case.

See microsoft/snmalloc#461.