Use the custom memcpy for realloc.
This wraps our memcpy with some assumptions that let the optimiser know
that we're copying chunks that are strongly aligned.  With clang 13 on
x86, this generates three variants:

 - A special case for 16 bytes that's a single vector load + store.
 - A vector-copy loop for sizes <512 bytes.
 - rep movsb for larger sizes.

This is almost certainly faster than the platform memcpy (if for no
other reason than that it doesn't have to care about handling unaligned
copies).

I don't know if it will show up in benchmarks, but if it does then it
Fixes #154