[RFC] Plumb additional argument through *Range types
We'd like our Metaslab*-s to be more tightly bounded than our chunks, but that
seems to mean that we need to split the backend allocation chain so that large
requests can enter "further along" than the SmallBuddyRange.

This is some "purely syntactic" noise to set the stage for doing that in the
next commit, factored out in an attempt to ease review.