Allow POSIX PALs to overwrite writev and fsync.
On FreeBSD (possibly elsewhere) the normal versions of these go via an
indirection layer because they are pthread cancellation points.  This
indirection layer does not work correctly if pthreads can't allocate
memory and so we can't get debug output until malloc is working, at
least a little bit.

With this version, we can call the __sys_ variants, which skip any libc
/ libthr interposition.