Remove indirection of state in ranges.
This doesn't give any extra flexibility: the range itself can be either
a stateless class, a class with no per-instance state that stores all of
static fields, or a class with stateful instances.  It did add a
requirement that every range implementation added an indirection layer.

Removing this removes over 100LoC with no loss of functionality.