Refill heuristic (#511)
* Adding a refilling heuristic

The large buddy allocator requests memory from its parent range.  The
request size was a fixed large request.  This was sufficiently large, so
that contention was not a problem.

This change makes it initially smaller, and gradually growing so that
contention is still not a problem, but for small work loads it requests
less memory.

* Remove special case for OE as no longer required.