Enable accessing parent ranges.
This exposes a feature on Ranges to access ranges higher up the
stack of ranges.  This could be useful for applying operations in the
middle of a pipeline like

   object_range.ancestor<SpecialRange>().init(...);

This allows some initialisation to be added to the middle of pipeline
without breaking the current coding pattern.

It also allows for bypassing some ranges

   object_range.ancestor<LargeObjectsRange>().alloc_chunk(...);

Neither are done in this commit, but both will occur in future commits.