In Checked mode use nullptr for malloc(0)
The spec allows malloc(0) to return an actual allocation, or nullptr.
Some code has compatibility issues with assuming if malloc returns nullptr
then the memory is full.  Hence, we normally use malloc(0) to return an
allocation.

In the checked client it seems better to return nullptr, and I believe
we have more opportunity for clients to find errors with it.