Removing benchmarking from PR triggers (#800)
The benchmarking workflow cannot run successfully on the PR as it
requires access to secrets which are not available in PR contexts.

This commit removes that trigger and adds a trigger on pushes to
branches starting with "bench/" so that testing can be done before
merging to main by those with write access.